/ init: history: sizevars: make empty -> unlimited
when we force the size to a large value (since there never used to be
"unlimited" support, it had served us well since olden days.

later bash versions actually try to allocate and fail (it could be libc
that doesn't like this -- the error is from xmalloc, failing on some
enormous number -- but bash 4.3 works and bash 4.4 does not with same
libc)

the documented way to make this unlimited is to use '-1' but this does
not work on older versions of bash.  empty string works for both values
(and this usage is documented for HISTFILESIZE)

before this patch, upgrade to bash 4.4 broke entire system, even root's
shell would not start since it was set to /bin/bash and had this large
value trick in its .bashrc