- alias: gcpinit: replace test path with standard
when we first installed google cloud tools it was a manual download and
install of the tools into a specific location for testing (named 'sdh'
due to the company name interviewing me at the time, ie my original
reason for learning google cloud tools)

since then, these tools are in a standard location and installed as an
os package, so move the path we source there to the standard place.
this way it works on any system that has the package installed.