- env: pdsh: don't try to force agent forwarding
agent forwarding is unsafe compared to stdio forwarded tunnel, it relies
on the security of the intermediary.  we should really never use it, but
at least require the config to come in from the outside, rather than
forcing it in our own config.  this may also have been causing some
interaction issues with controlpersist multiplexing