^ funcs: remove google commands from personal, mv to work
work should have these and we don't want to force people there to source
files, so they're put into a script as functions with symlinks to the
script and switch on argv0.

this means we can't really share them for personal, unless we have that
dir in path.  maybe we'll do that, not really a good solution, but since
it's the "right way" for work we'll do that for now and figure out the
personal stuff later (just don't want to have it in two places, so
removing it from here while we're adding it to the other place)