- funcs: grepp: make the list+color variants actually invoke color
this aliases the *l variants to invoke the non-l variants for the c*
case, because c*l makes no sense: it doesn't show any contents, so
there's nothing to colorize.  but when working up a command line by
iteration, we often start with the list function, narrow down the
pattern, and then we want to switch to non-list (show contents) and
colorize.  so we just want to add a 'c' at the beginning and not have to
remove the 'l', that requires too many keypresses