+ env: fontconfig: path: specify for gui apps that need it
we also set this in windowmanager because some things launch from there
without use of the shell.  this is required because apparently the
hardcoded path in fontconfig is that of the build system used to create
and ship it, instead of something sane.  if we don't do this things like
chrome will complain that no path is set