- funcs: taskw: task: special-case --version without usual args
taskwarrior doesn't seem to give version info for example if you give it
with rc.defaultwidth along with --version.  so we'll just give *only*
version info, if we see --version anywhere on the command line (prior to
a '--' end-of-arguments indicator).  can't think of a situation where
this would cause a problem