+ alias: alltimes: list all timew/ prefixed timewarrior tasks
for us, not all timew tasks correspond to taskwarrior tasks.  we use the
'time/' namespace for these.  they are just things that have to do with
timewarrior itself or are just "time sinks" that are recurring

this alias is handy to quickly list them all by frequency, for deciding
which one to use