+ func: taskw: retimew: make sure label looks up before applying
both old and new labels should look up correctly, otherwise we should
abort before we start.  this can avoid confusion about usage, such as
right now when trying to retimew app/ to src/ but actually this isn't a
label it's just a component and that's not what this function does.
such an operation still needs to be done manually