+ init: term: disable xon/xoff which we don't use anymore
not really sure why this isn't the default for the input line
discipline.  maybe it can be set with some kernel parameters or
something.  but we'll just set it in our shell for now.

we're doing this unconditionally so hopefully it works on all systems...
at least we'll squelch stderr in case it does nothing or doesn't exist,
it shouldn't really be a problem, and term.d/ are only executed for
interactive terminals