+ func: taskw: timesince: make timew range from relative duration
make a range "from - to" with those being iso8601 datetimes, so the arg
is just the duration and then it will start that far back in time and
end now.  OR it will take a second arg of start time and have it be
relative to *that* time as the start.