+ func: taskw: taskrecent: add stat chars: completed, started, todo
this lets us see at a glace what we're working with, in practice this
will be highly useful because otherwise we have to think about each one
and say "is that done already? was it already started or just entered?"

todo: this only sorts based on last-modified time for tasks, which can
be disconnected entirely from last-worked-on tasks.  however, if we use
timew output, we don't have status information, so we would need both
and that would end up being a lot more complicated.  in practice we can
probably just look at taskstat for the latter...