/ func: taskw: taskcur: only use exact taskid from tasknow result
taskid usually gives the first match, which isn't always the exact
match.  this means that, even though we know the exact current fql (from
tasknow, which uses timew data), when we try to get the taskid from it,
we might get a different result than the actual current task (example,
if fql is a substring of another task that has same label initial
substring and in same project).

since tasknow always gives the complete fql, it's already unique and
taskid -xn should always work and always be exactly one match and always
be the correct match