+ func: history: lhistory: use real max int by runtime calculation
this works on any system

two's complement system, 64 bits.  leftmost bit is sign bit.  bits after
it represent an absolute number to be added to negative maxint,
resulting in the negative number it encodes for.

 $ qalc -s hextwos -s "bits 64" "-1 to hex"
 −1 = 0xFFFFFFFFFFFFFFFF

then change the first bit (only) to a zero, which means unsigned, which
means the bits just represent themselves in two's complement.  in other
words, the max positive number:

 $ qalc -b 2 -s hextwos -s "bits 64" $(
     qalc -s hextwos -s "bits 64" "-1 to hex" | field -1 | sed s,F,7,)
 9223372036854775807 = 0111 1111 1111 1111 1111 1111 1111 1111
                       1111 1111 1111 1111 1111 1111 1111 1111