/ env: less: use $LESSOPEN, not $LESSPIPE, must be old or mistake
cannot find reference to LESSPIPE anywhere, this is either old or an
error.  it's possible that older LESS only supported use as a pipe, but
less of today can support using a temporary file for example.  the
typical usage is to call a program called "lesspipe" though, so maybe
this was just an error.  we don't actually use it anyways but probably
did on an older system...