- env: less: move $LESS to be next to $PAGER, which uses 'less'
should be clear that one is using the other by having them be next to
each other.  ok to leave LESSOPEN somewhere else because it has a more
complicated initialization that tests for presence of lesspipe.sh and
requires a comment.  especially ok since we're not actually using
lesspipe.sh anyways at the moment.