^ env: less: remove '-n' which suppressed line numbers
there was a time when machines were slow that not writing line numbers
may have been a useful optimization.  vaguely recall this causing a
problem on some old slow machine, probably when looking at huge log
files.  it would pause and say "calculating line numbers..." when
seeking to some part of the file using eg the '%' key.

however, after looking at the source blame data, it seems that a sparse
linenum cache (ie, every so often, offset/linenum correlations are
recorded, and then they can be recalculated relative to the closest
cached one, to zero in on a particular line), has been implemented since
at least 1994, and that goes back to my early days with linux (before
that, would have only encountered 'more'), so it's doubtful that the
cache was not in place when encountering the slowness before.

one bonus with line numbers is that using 'v' to callout to $EDITOR will
add +n (if vi-like) and jump to that offset.  we do use that feature on
occasion.

in any case, cannot imagine it being terribly slow on any machines we
are using today... and we can always add '-n' manually to invocation
args if that comes about, or change the default back, if/when any sort
of delay is encountered again.  hopefully will remember this commit!