+ env: less: add '-F' to $LESS so less than a screenful exits
this way it's like we're not using a pager at all for stuff less than a
screenful of output.  and bonus, it won't use secondary screen either.
so it's almost like the 'less' didn't exist at all, which is perfect for
most use cases.

we had found this before and added it manually for some stuff, but no
reason this can't be automatic for all invocations.  can't think of a
situation where it wouldn't be wanted...