^ env: elinks: remove $ELINKS_CONFDIR, we have to use symlink
when invoked from xdg-open, it's using a clean environment so this env
var does not carry over.  we have to use the symlink method instead
.config/elinks/ -> ../.elinks/

this is too bad because this symlink can't be versioned using any
existing methods.  but elinks author didn't like keeping compatibility
for some reason, see upstream issue 199.