- func: list: tree: switch from ansi-escaped to utf8 linedraw
VT100 can do line drawing by escapes, there is an alternate graphics
character set one can switch to with 'escape-(N' where N is 0-4, and
leave it with 'escape-(B'.  for example:

    echo -e '\e(0q\e(B'

will draw a horizontal line character and then leave the mode.

that is the method employed by 'tree' when using '-A' for line drawing.
this, in turn causes issues with 'less -R' because these escapes will be
stripped as they are not strictly color escapes.  instead we would need
to use 'less -r' which opens us up to other issues such as embedded
malice in the filenames being listed by 'tree', and makes horizontal
scrollling problematic because less cannot correctly count the
characters.

however, it turns out the 'tree' actually can output real utf8 line draw
characters; in fact it seems to do this by default, which must have
changed at some point (it used to show ASCII characters for line draws,
which was the whole reason for using -A in the first place), but there
might still be circumstances where it will output ASCII.

to be explicit, we will force the matter by forcing utf8 character set.

this was all learned by reading the source code, which was itself hard
to find: https://github.com/Old-Man-Programmer/tree/

VT100 line draw escapes are documented here:
https://en.wikipedia.org/wiki/DEC_Special_Graphics