+ func: taskw: task: add back just task(), it belongs in personal rc
this indirection for 'task' invocations is only doing two things:

    - if --version requested, just give that and exit
    - otherwise, pass along args, but set rc.defaultwidth=$COLUMNS
        - helps with piping output of 'task' to things

originally we had moved the function into taskwtools along with all the
others in "the big move" from bash profile functions to our taskwtools
task.sh invname-switcher, but it's causing problems there, and adding
wrap in a pipeline might be said to be a local decision anyways.  it
seems we can't reliably determine the full path to a script accessed
through a symlink, without calling readlink, anyways.

now, we will move it back into our bash profile, for the moment, it
solves a lot of problems if taskwtools does not have to make this work
in a standalone script