/ rc: path: ~/bin/ should be in the non-interactive path
in 70d1c56b we added /usr/local/ bins to the non-interactive path

here we complete that by doing it also for the homedir bin.  otherwise
we have to make special arrangements when we ssh in and want to utilize
the scripts in bin/ indirectly (such as git <script> will fail unless
PATH is set right)

we didn't add ~/.local/bin/ because that will be going away and we will
use symlinks.  do not want to have one bin per language.  still thinking
about this, however...