+ env: less: add --no-vbell as it just happens too often
nearly everthing in the UI that it doesn't like will ring your terminal.
we told it not to with -Q, but that only makes them visual.  now kill
those too... generally we know some keypress was wrong/invalid because
no action occurred in response to the keypress.  and we've been using
less for 30+ years now