+ alias: pdshx: skip pdsh_all_skip group without specifying every time
annoying to have to keep typing this, we use it often with nodes that
aren't always up.  when using '-a' it's not needed but other invocations
don't pick it up.  they probably should, maybe we should make a patch
upstream... TODO

oh and added a root variant also since that's another common thing