/ completions: incus: gate on presence of ~/.config/incus
already we gate on incus existing, but we should also gate on whether
incus has been run before.  this directory is created on incus init,
there is some kind of news item that runs for first time when we service
incus completions that gives an operational blurb to the user.  we don't
want that to happen and no need to load completions on systems where we
have never run incus and will likely not run it (always can
re-initialize if needed after the fact)