/ promptcmd: winch: actually send the window change signal
SIGWINCH was never actually getting sent because the trap was always
installed, so even raising it inside the trap would just enter the trap
again.  what a silly oversight... that means apparently SIGWINCH
handling is unnecessary for ordinary operations, since it's been like
this since 9c4c66ba, which was in 2016.  it does mean SIGWINCH was
effectively ignored, so whatever that did must have helped at the time
because it fixed the bug it addressed.  but, really we should still send
a SIGWINCH, just do it at the right time.  it's possible this explains
some anomalies seen before with prompt text filled out but the window
resizing before command has been fully entered.  maybe... but in any
case now we'll try to do it the right way at least, going forward.