/ rc: stash cwd before changing dirs then go back after init
we change dirs during shell rcfile initialization for various globbing
and other purposes, but then we just blindly chdir to the home dir at
the end.  instead we'll save the cwd early and restore it at the end, so
we can start another shell level anywhere and not have to "cd" to the
dir we started it in.  really this is a bug, don't think it was meant to
be like this, but it doesn't come up often because we like the default
of being in the home dir.  but we were only doing that because more
times than not that's where it started out (ie, login shell).