+ promptcmd: add command start time and duration to logged values
for a long time we have been putting the start time in the bash history
file, and the prompt generation time (which is the end of the previous
command) in cmdhist.  now we move the start time calculation earlier so
we can put both values -- and then precompute the duration -- inside the
cmdhist file.  this is important information.  the duration could be
computed of course by viewing software but it will be distributed over
much more time if we precalculate it with each history line addition