+ history: log also to sqlite3 file with same basename as cmdhist
we will start logging all our commands and extra metadata to sqlite3
tables inline.  eventually we'll collate the ones from all the hosts and
have a central place for them and hook into bash with new loadable
builtin modules to access the history data and export keybinding
commands to libreadline so we can do hist init in bash directly from the
sql file and have it updated in-memory in realtime.