/ promptcmd: ensure command line has nonzero length before logging
for some reason, from somewhere, some of these are slipping through with
zero length, not sure where these are coming from since we have the
other checks, but they are fairly rare.  caught it when importing to
sql, they are only a few dozens out of millions of lines, but they exist
and some are relatively recent, only a few thousand commands ago.
probably there is some rare occasion when it happens.  it will be
discarded now as a log event, since it contains no command.