/ promptcmd: use recent background pid $! to avoid race condition
when we just use no argument, it will use the default, which is like
using the job specifier %%, ie the "current job," which is sometimes
different than the process whose pid has a value of $!, because remember
it's in the shell context already and for example if we have just added
a job like "man bash" and then control-z, most recently backgrounded
will be different than %% which was not yet backgrounded.  this means if
we use %% in prompt_command() it will refer to the stopped bash, whereas
$! will refer to the command before the backgrounding ampersand.

in theory a race still exists... we could background something while
prompt_command is still in the middle of the sqlite3 but before the
disown was successful and returned.  this would require very fast typing
because the command is issued before the user inputs any text! maybe
with a paste, but even then, the window is so small, it seems it would
be impossible in practice.  we'll see if this actually comes up in
practice, very doubtful.