+ completions: hide expensive completions behind a "completions" function
adding the linode-cli completions, it became clear these were just too
expensive to always add.  so we put the more expensive completions
(especially the ones that call some other program to generate
completions on its stdout) behind a variable and only set it when
sourcing the completion file via 'completions' shell function.  this way
the user has to be explicit that they want to do the expensive
completion generation stuff specifically for this shell.  since it's
pretty uncommon to do it generally, and only a specific shell will
typically be used for it, this should help the fast path of new terminal
window substantially.