/ promptcmd: sqlite3: retry for 3s as we sometimes race other shells
sometimes we get "Error: near line 2: in prepare, database is locked (5)"
when a long-running shell exits asynchronously and we happen to be
running a command at the same time.  adding an internal
retry-until-timeout should take care of it.  using WAL mode might also
work, but that will mean additional files in the logdir which we want to
avoid for now.