^ env: termux: gate termux vars on just prefix var, not termux version
when we enter termux environment by adb, our script doesn't hardcode
TERMUX_VERSION and it doesn't get set automatically for us, however we
*are* willing to hardcode TERMUX__PREFIX because that should never
change.  really looking through all the termux code and scripts has made
me realize this is a pretty gigantic hack and we should not get too deep
into customizing or using termux, we should only use it in a limited way
because it's just too different and modified from standard linux.

in any case we'll gate on the prefix var only for now and leave the
previous extra version var commented out, maybe remove later or consider
re-adding it