+ history: allow hint not to save command if line starts with space
allow user to hint not to save it by starting line with a space.  we
will use this in particular for task annotations and other.  this might
miss some important things where we accidentally have a space, but we'll
just have to be aware of that... usually we're pretty good with this
kind of thing