^ mkprompt: hostname: unnecessary to set this, should always be set
there should be no reason to set this as the shell does it itself
according to the manual page.  nonetheless we leave a guard in there to
make sure.  this doesn't actually matter since mkprompt is run only if
the prompt input file has changed so isn't performance critical, but
still there's no reason to do it.

also note that strangely, while it gets set by the base shell, it
doesn't seem to be exported on its own, so we do leave that in place.