/ agents: create the tmpdir in home, ~/tmp if it does not exist
we recently changed to use ~/tmp/ instead of /tmp/, but this doesn't
always exist so we do have to create it.  maybe this should be done
elsewhere in a higher level init but so far this is the only place using
it and the os will cache the result and it's a builtin so shouldn't
affect speed much to do simple "test -e".  too bad "mkdir -p" doesn't
use a builtin...