- aliases: tz: switch to the canonical versions of time zones
so we don't rely on tzdata-legacy packages, on u24 these time zones were
put in an optional package and deprecated, apparently it's been this way
for years in the iana data.  the rationale is that the city will always
be correct, but its time could change, for example if a city decides on
UTC, which seems unlikely in the case of America ;-) Nonetheless, we
don't want this extra package requirement

we have always preferred actually the EST5EDT for example as these are
more descriptive, but they have the same failing in that the city can
change what time zone they use and/or lines could be redrawn.