- completions: kubectl: also explicitly complete kubectl itself
even though this repeats what was already done as a final step in the
"kubectl completion bash" script, we want to be explicit and if the 'k'
alias breaks we want 'kubectl' to break also in the same way to prevent
confusion.  anyways it must use a space because that's what the
auto-generator says to do for the default kubectl completion.