+ func: k: wrap kubectl with namespace+config and completions
we don't use context because that requires modifying keys in a yaml
config file.  this (1) affects things for all user and all shells and
(2) requires generating a transient config file because we want an
unmodified /etc/kubernetes/admin.conf from the primary master node, in
order to do this while specifying context we need to merge them on the
fly.  using command-line arg wrapping is also hackish, especially in a
shell function, but this was least effort for now and see where this
goes.  there is also kubeswitch and kubectx/kubens but those seemed
unnecessarily complicated for our particular situation, which is just
switching plex.  once we have many clusters, may revisit this.  then
again, each one having a config file, and persistent namespace, can
scale pretty good maybe.