+ env: systemd: pager: do not use a pager with systemd and force colors
have the outside user use the pager, but default to complete output
every time.  this way journcalctl and systemctl status is more like
"logcat" and user has to be aware when output might be copious.  a
difference in behavior that we'll have to retrain for, but suspending
journalctl with ctrl-z and coming back to is is very frequent and it has
to work.