+ completions: cilium: add completions to our plex kube cni tool
this is useful to have like all the others.  as before everything is
gated on having run 'completions' function so most shells don't carry
the weight, just ones where we want to.  TODO: these functions all
follow the same pattern and repeat the same tests.  should factor out
the parts that are the same into multi-level conditionals.