+ completions: gh: add completions for github cli commands
this will be useful for sure, in certain shells.

there's starting to be a lot of completions, maybe we need our
completions() function to take an argument to group them... not all
shells would want all completions, rather than the current all-or-none
(default none).  for the most part though, maybe the same shell would
want them all, we'll see what happens in practice.